requirements.txt
apache-airflow==2.10.5
requests == 2.32.3
kafka-python == 2.1.2
